==> 01_Combine_orginal_datasets.ipynb <==
import pandas as pd

# List the names of the CSV files
csv_files = ['April12_2020.csv','Jan12_2021.csv','April28_2021.csv','Jan24_2022.csv']

# Create an empty list to store the DataFrames
df_list = []

# Iterate over the CSV files
for file in csv_files:
    # Read the CSV file into a DataFrame
    df = pd.read_csv(file)
    # Add the DataFrame to the list
    df_list.append(df)

# Concatenate the DataFrames into a single DataFrame
combined_df = pd.concat(df_list)

# Write the combined DataFrame to a CSV file
combined_df.to_csv('combined_final_orginal_dataset.csv', index=False)

import pandas as pd

# Read the Excel file into a DataFrame
df = pd.read_csv('combined_final_orginal_dataset.csv')

# Get the number of rows in the DataFrame
num_rows = df.shape[0]

# Print the number of rows
print(num_rows)

df.head()
