==> 02_Text_cleaning.ipynb <==
import pickle
import pandas as pd
import re
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import train_test_split
from sklearn.feature_selection import chi2
import numpy as np
import string

f=pd.read_csv("combined_final_orginal_dataset.csv")

#Remove unneccasry columns and create a new csv

keep_col = ['created_at','lang','text']
new_f = f[keep_col]
new_f.to_csv("Updated_Orginal.csv", index=False)

#Remove non english

new_f=pd.read_csv("Updated_Orginal.csv")
new_f = new_f[new_f["lang"].str.contains("en") == True]
new_f.to_csv("Updated_Orginal_Eng.csv", index=False)

num_rows = new_f.shape[0]
print(num_rows)

#remove the duplicate rows based on the 'text' column

new_f=pd.read_csv("Updated_Orginal_Eng.csv")

# Remove duplicate rows based on the 'text' column
new_f = new_f.drop_duplicates(subset =['created_at','lang','text'], keep='first', inplace=False, ignore_index=False)

new_f.to_csv('Updated_Distinct_Orginal.csv', index=False)


num_rows = new_f.shape[0]
print(num_rows)

Data Preprocessing : Data Cleaning

#Remove http web references

new_f['Content_Parsed_0'] = new_f['text'].str.replace('https:\/\/.*|http:\/\/.*',"")

#Remove emails @

new_f['Content_Parsed_1'] = new_f['Content_Parsed_0'].str.replace(r"\S*@\S*\s?", "")

#Remove possessive pronouns

new_f['Content_Parsed_2'] = new_f['Content_Parsed_1'].str.replace("'s", "")

#Remove special characters

new_f['Content_Parsed_3'] = new_f['Content_Parsed_2'].str.replace("\r", "")
new_f['Content_Parsed_3'] = new_f['Content_Parsed_3'].str.replace("\n ", "")
new_f['Content_Parsed_3'] = new_f['Content_Parsed_3'].str.replace("\n", "")
new_f['Content_Parsed_3'] = new_f['Content_Parsed_3'].str.replace("    ", "")

new_f['Content_Parsed_4'] = new_f['Content_Parsed_3'].str.replace("&", "")
new_f['Content_Parsed_4'] = new_f['Content_Parsed_4'].str.replace("#", "")

#Lowercasing the text

new_f['Content_Parsed_5'] = new_f['Content_Parsed_4'].str.lower()

#Remove Punctuation signs

#.replace({"$^'|'": ""}, regex=True)
punctuation_signs = list("!#$%&'""()*+,-./:;<=>?@[\]^_`{|}~")
new_f['Content_Parsed_6'] = new_f['Content_Parsed_5']

for punct_sign in punctuation_signs:
    new_f['Content_Parsed_6'] = new_f['Content_Parsed_6'].str.replace(punct_sign, " ")
    
new_f['Content_Parsed_6'] = new_f['Content_Parsed_6'].str.replace('"', '')
new_f['Content_Parsed_6'] = new_f['Content_Parsed_6'].replace({"$^'|'": ""}, regex=True)
new_f['Content_Parsed_6'] = new_f['Content_Parsed_6'].str.replace('’','')
new_f['Content_Parsed_6'] = new_f['Content_Parsed_6'].str.replace('“','')
new_f['Content_Parsed_6'] = new_f['Content_Parsed_6'].str.replace('”','')

#Remove numbers

new_f['Content_Parsed_7'] = new_f['Content_Parsed_6'].str.replace("\d+", "")

#Remove all single characters

new_f['Content_Parsed_8'] = new_f['Content_Parsed_7'].str.replace(r'\b\w\b','').str.replace(r'\s+', ' ')

#Remove whitespaces

new_f['Content_Parsed_9'] = new_f['Content_Parsed_8'].str.replace(r"^\s+","")
new_f['Content_Parsed_9'] = new_f['Content_Parsed_8'].str.replace(r"\s+"," ")

new_f.to_csv('pre_processed_data.csv')

#orginal text
new_f.loc[1]['text']

#remove http
new_f.loc[1]['Content_Parsed_0']

#remove email @
new_f.loc[1]['Content_Parsed_1']

#remove posessive pronoun
new_f.loc[1]['Content_Parsed_2']

#remove special characters
new_f.loc[1]['Content_Parsed_3']

#remove special characters
new_f.loc[1]['Content_Parsed_4']

#lowercase
new_f.loc[1]['Content_Parsed_5']

#remove punctuation signs
new_f.loc[1]['Content_Parsed_6']

#Remove numbers
new_f.loc[1]['Content_Parsed_7']

# remove all single characters 
new_f.loc[1]['Content_Parsed_8']

#remove whitespaces
new_f.loc[1]['Content_Parsed_9']

new_f.tail()
