==> 03_TokenizationLemmatization.ipynb <==
import pickle
import pandas as pd
import re
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import train_test_split
from sklearn.feature_selection import chi2
import numpy as np

new_f=pd.read_csv("pre_processed_data.csv")

new_f.tail()

new_f.loc[132266]['Content_Parsed_9']

Tokenization

from nltk.tokenize import word_tokenize

new_f['Content_Parsed_9'] = new_f['Content_Parsed_9'].apply(str)

# Tokenize the text in the 'text' column
new_f['Content_Parsed_10'] = new_f['Content_Parsed_9'].apply(nltk.word_tokenize)

new_f.loc[2]['Content_Parsed_10']

!pip install nltk
!pip install wordnet


POS Tagging and Lemmatization

from nltk.stem import WordNetLemmatizer
from nltk.corpus import wordnet
from nltk.tokenize import word_tokenize
from nltk.tag import pos_tag

import nltk
nltk.download('averaged_perceptron_tagger')

import nltk
from nltk.corpus import wordnet
from nltk.stem import WordNetLemmatizer

def get_wordnet_pos(treebank_tag):
    if treebank_tag.startswith('J'):
        return wordnet.ADJ
    elif treebank_tag.startswith('V'):
        return wordnet.VERB
    elif treebank_tag.startswith('N'):
        return wordnet.NOUN
    elif treebank_tag.startswith('R'):
        return wordnet.ADV
    else:
        return wordnet.NOUN  # default to noun

nrows = len(new_f)
lemmatized_text_list = []

for row in range(0, nrows):
    
    # Save the text and its words into an object
    tokens = new_f.loc[row]['Content_Parsed_10']
    pos_tags = nltk.pos_tag(tokens)
    
    lemmatizer = WordNetLemmatizer()
    lemmas = []
    for token, pos in pos_tags:
        # Determine the part of speech of the token
        pos = get_wordnet_pos(pos)
        # Lemmatize the token
        lemma = lemmatizer.lemmatize(token, pos)
        lemmas.append(lemma)
        
    lemmatized_text = " ".join(lemmas)
    
    # Append to the list containing the texts
    lemmatized_text_list.append(lemmatized_text)
    


new_f['Content_Parsed_11'] = lemmatized_text_list

new_f.loc[79074]['Content_Parsed_11']

#Stop words Removal

# Downloading the stop words list
nltk.download('stopwords')

# Loading the stop words in english
stop_words = list(stopwords.words('english'))

#remove words that are in NLTK stopwords list
not_stopwords = {'not'} 
updated_stopwords = set([word for word in stop_words if word not in not_stopwords])

nrows = len(new_f)
filtered_words = []

for row in range(0, nrows):
    s_list = []
    
    # Save the text and its words into an object
    text = new_f.loc[row]['Content_Parsed_11']
    text_words = str(text).split(" ")
    for word in text_words:
    # If the word is not a stop word, add it to the filtered list
        if word not in updated_stopwords:
            s_list.append(word)
        
    # Join the list
    stop_text = " ".join(s_list)
    
    # Append to the list containing the texts
    filtered_words.append(stop_text)
    

new_f['Content_Parsed_12'] = filtered_words

#consecutive emoji seperation

nrows = len(new_f)
text_list = []

for row in range(0, nrows):
    
    # Save the text and its words into an object
    text = new_f.loc[row]['Content_Parsed_12']
    words_and_emojis = re.split(r'([\U0001f600-\U0001f64f\U0001f300-\U0001f5ff\U0001f680-\U0001f6ff\U0001f1e0-\U0001f1ff\U00002702-\U000027b0])', str(text))
    
    elements = []
    for element in words_and_emojis:
        if re.search(r'([\U0001f600-\U0001f64f\U0001f300-\U0001f5ff\U0001f680-\U0001f6ff\U0001f1e0-\U0001f1ff\U00002702-\U000027b0])', element):
            for emoji in element:
                elements.append(emoji)  
        else:
                elements.append(element)
    
    emo_text = " ".join(elements)
    text_list.append(emo_text)


new_f['Content_Parsed_13'] = text_list

nrows = len(new_f)
text_list = []

for row in range(0, nrows):
    text = new_f.loc[row]['Content_Parsed_13']
    new_list = text.split()
    
    text_list.append(new_list)
    
new_f['Content_Parsed_13'] = text_list

new_f.loc[4]['Content_Parsed_13']

#Remove duplication

import ast

def remove_duplicate_words(lemmas):
  # Create a new list of unique lemmas
  unique_lemmas = []
  for lemma in lemmas:
    if lemma not in unique_lemmas:
      unique_lemmas.append(lemma)
  return unique_lemmas

nrows = len(new_f)
text_list = []

for row in range(0, nrows):
    
    word_list = []
    
    # Save the text and its words into an object
    text = new_f.loc[row]['Content_Parsed_13']
    text_words = str(text).split(" ")
    
    list_text = " ".join(text_words)
    
    # Append to the list containing the texts
    text_list.append(list_text)
    
for i, element in enumerate(text_list):
    # Apply ast.literal_eval() to the element
    element = ast.literal_eval(element)
    # Replace the element in the list with the evaluated value
    text_list[i] = element
    
new_f['Content_Parsed_14'] = text_list

new_f['Content_Parsed_14'] = new_f['Content_Parsed_14'].apply(remove_duplicate_words)

#Remove Emojis and Replace it with its meaning as a text

from emot.emo_unicode import UNICODE_EMOJI # For emojis
from emot.emo_unicode import EMOTICONS_EMO # For EMOTICONS

def convert_emojis(text):
    for emot in UNICODE_EMOJI:
        text = str(text).replace(emot, "_".join(UNICODE_EMOJI[emot].replace(",","").replace(":","").split()))
    return text


nrows = len(new_f)
emoji_text_list = []

for row in range(0, nrows):
    emo_list = []
    text1 = new_f.iloc[row]['Content_Parsed_14']
    textEmo = convert_emojis(text1)
    emoji_text_list.append(textEmo)
    

new_f['Content_Parsed_15'] = emoji_text_list

words = set(nltk.corpus.words.words())

f = lambda x: " ".join(w for w in nltk.wordpunct_tokenize(x) if w.lower() in words)
new_f['Content_Parsed_16'] = new_f['Content_Parsed_15'].apply(f)

Outputs

#cleaned text
new_f.loc[79074]['Content_Parsed_9']

# Tokenization
new_f.loc[79074]['Content_Parsed_10']

#POS and Lemmatization
new_f.loc[79074]['Content_Parsed_11']

#stop word
new_f.loc[79074]['Content_Parsed_12']

#seperate emojies
new_f.loc[79074]['Content_Parsed_13']

#Remove duplications
new_f.loc[79074]['Content_Parsed_14']

#replace emoji with word
new_f.loc[79074]['Content_Parsed_15']

main_columns =["created_at","text","Content_Parsed_15"]
new_f=new_f[main_columns]

new_f = new_f.rename(columns={'created_at':'Date','text':'Orginal tweet','Content_Parsed_15':'Preprocessed tweet'})

new_f.head()

new_f.to_csv('Preprocessed_columns.csv')
